requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
aioredis
pydantic
python-dotenv
geopy
pandas
scikit-learn
numpy